backend/requirements.txt
fastapi
uvicorn
pydantic
pandas
numpy
faker
sqlalchemy
openai
python-dotenv
